my_diplom/requirements.txt
selenium
webdriver-manager
requests
pytest
python-decouple
allure